This should fix the issue with pixels per point not properly being set by the UI. This allows the individual control nodes to properly scale egui as needed.
Also made some minor changes to the function call of paint_shapes, since it is already using &mut self, rather than passing the texture into the method, the method queries for it directly.

Also added some additional widgets to the example to test the mouse position is correct regardless of the pixels_per_point setting.