Finished up the first pass on the color and demo integrations.
This integrates the existing work as well as creates a shader at compile time so that there's no additional configuration for the egui node.
The color tests don't fully pass, this is likely due to a blend mode issue.
TODO: Include an FAQ to explain certain smaller issues related to color.